fix, rename iCals to cals
